Assets
I used the stretched out Peter Griffin for the pipes, Joe Swanson's cutout face as the Flappy Bird stand-in, Sam Neil in "In the Mouth of Madness" theater scene as the background, the song "The Penis-Eek" as the background music, and that dude doing the really loud AAAUUUGGGH as the game over sound. Feel free to use different assets in your own rendition.